feat(investment): ADR-090 per-config property selection pipelines (Phases A–E)
Implements independent selection pipelines per frontier configuration,
ensuring each config draws from a genuinely distinct property cohort.

Phase A — `FilterAffordable` pre-screen: filters property pool to budget-
qualified candidates before cohort selection; replaces post-hoc elimination.

Phase B — `PropertyCohort` type + three named pipelines:
- `SelectQualityCohort` (Quality): AI score + cap rate composite
- `SelectIncomeCohort` (Income): cap rate + CoC composite
- `SelectBalancedCohort` (Balanced): overlap-penalised mixture
- `BuildCohorts` orchestrator dispatches all three, measures Jaccard < 0.40

Phase C — `Recalculate` fast path: re-evaluates Markowitz objectives from
property data on every call (appreciation rate + NOI formula via
`ExpenseRatioForAge`), ensuring sliders always produce consistent updates.

Phase D — `GenerateLabel` (exported): human-readable label matrix covering
all configType × strategy × risk combinations; flows through to `FrontierPoint`.

Phase E — `SelectGrowthCohort` for Appreciation strategy:
- Formula: 0.4×normPrice + 0.35×normAI + 0.25×normBudgetCapacity
- Replaces Quality cohort when strategy == "appreciation" (3 cohorts total)
- `ConfigType` promoted from string to named type with constants
- `computeNormsWithBudget` with div-by-zero guard for BudgetCapacity

Also: `AcquisitionPricer` wired with real market data via MetroReader;
fallback to US national medians ($425K price, $1,987 rent). PDF rewritten
to mirror workspace screen order. Fix `TestRecalculate_NoOverrides` and
`TestCreateCohort` to match new re-evaluation behaviour.

+++ internal/services/investment/projection/calculator.go
@@ -794,8 +794,8 @@ func (c *Calculator) CalculateCombinedMetrics(
 
 		// Combined metrics
 		CombinedPropertyCount:  existing.PropertyCount + newMetrics.PropertyCount,
-		CombinedTotalValue:     existing.TotalValue + newMetrics.TotalInvestment,
-		CombinedAnnualCashFlow: existing.AnnualCashFlow + newMetrics.AnnualCashFlow,
+		CombinedTotalValue:     existing.TotalValue + float64(newMetrics.TotalInvestment),
+		CombinedAnnualCashFlow: existing.AnnualCashFlow + float64(newMetrics.AnnualCashFlow),
 	}
 
 	// Calculate improvements
